# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.RegisterSocialMediaPayload:
    properties:
      link:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  main.RegisterWorkingAtPayload:
    properties:
      company:
        type: string
      github:
        type: string
      instagram:
        type: string
      linkedin:
        type: string
      month:
        type: integer
      title:
        type: string
      totalyear:
        type: integer
    required:
    - title
    type: object
...
  main.UpdateSocialMediaPayload:
    properties:
      link:
        type: string
      name:
        type: string
    type: object
  main.UpdateWorkingAtPayload:
    properties:
      company:
        type: string
      github:
        type: string
      instagram:
        type: string
      linkedin:
        type: string
      month:
        type: integer
      title:
        type: string
      totalyear:
        type: integer
    type: object
...
  store.SocialMedia:
    properties:
      id:
        type: integer
      link:
        type: string
      name:
        type: string
      userid:
        type: integer
    type: object
...
  store.WorkingAt:
    properties:
      company:
        type: string
      github:
        type: string
      id:
        type: integer
      instagram:
        type: string
      linkedin:
        type: string
      month:
        type: integer
      title:
        type: string
      totalyear:
        type: integer
      userid:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Appointr server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Appointr API
  version: "1.0"
paths:
...
  /socialmedia/{socialMediaID}:
    delete:
      consumes:
      - application/json
      description: Delete socialmedia by id
      parameters:
      - description: SocialMedia ID
        in: path
        name: socialMediaID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete socialmedia
      tags:
      - socialmedia
    get:
      consumes:
      - application/json
      description: Get socialmedia by ID
      parameters:
      - description: SocialMedia ID
        in: path
        name: socialMediaID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.SocialMedia'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get socialmedia by ID
      tags:
      - socialmedia
    patch:
      consumes:
      - application/json
      description: Update socialmedia
      parameters:
      - description: SocialMedia ID
        in: path
        name: socialMediaID
        required: true
        type: integer
      - description: SocialMedia
        in: body
        name: socialmedia
        required: true
        schema:
          $ref: '#/definitions/main.UpdateSocialMediaPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.SocialMedia'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update socialmedia
      tags:
      - socialmedia
  /socialmedia/create:
    post:
      consumes:
      - application/json
      description: Create a new socialmedia
      parameters:
      - description: SocialMedia
        in: body
        name: socialmedia
        required: true
        schema:
          $ref: '#/definitions/main.RegisterSocialMediaPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new socialmedia
      tags:
      - socialmedia
  /socialmedia/u/{id}:
    get:
      consumes:
      - application/json
      description: Get all socialmedia by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.SocialMedia'
            type: array
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all socialmedia by user ID
      tags:
      - socialmedia
...
  /workingat/{workingatID}:
    delete:
      consumes:
      - application/json
      description: Delete workingat by id
      parameters:
      - description: WorkingAt ID
        in: path
        name: workingatID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete workingat
      tags:
      - workingat
    get:
      consumes:
      - application/json
      description: Get workingat by ID
      parameters:
      - description: WorkingAt ID
        in: path
        name: workingatID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.WorkingAt'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get workingat by ID
      tags:
      - workingat
    patch:
      consumes:
      - application/json
      description: Update workingat
      parameters:
      - description: WorkingAt ID
        in: path
        name: workingatID
        required: true
        type: integer
      - description: WorkingAt
        in: body
        name: workingat
        required: true
        schema:
          $ref: '#/definitions/main.UpdateWorkingAtPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.WorkingAt'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update workingat
      tags:
      - workingat
  /workingat/create:
    post:
      consumes:
      - application/json
      description: Create a new workingat
      parameters:
      - description: WorkingAt
        in: body
        name: workingat
        required: true
        schema:
          $ref: '#/definitions/main.RegisterWorkingAtPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new workingat
      tags:
      - workingat
  /workingat/u/{id}:
    get:
      consumes:
      - application/json
      description: Get all workingat by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.WorkingAt'
            type: array
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all workingat by user ID
      tags:
      - workingat
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"